Rename files to full test names
